(Refact) extract Kakao login logic into shared utility function
